requirements.txt
numpy==2.3.3
pvporcupine==3.0.5
pyalsaaudio==0.11.0
